feat(recording): add device validation and fallback for VAD recorder
- Extract shared device validation logic into device-stream.ts module
- Add stream management and device fallback to VAD service
- Handle device acquisition outcomes in UI with informative toasts
- Standardize file naming: vad.ts → vad-recorder.ts
- Ensure consistent device handling across manual and VAD recorders

When a selected device is unavailable, both recording modes now:
- Automatically fallback to first available device
- Update settings with the fallback device ID
- Show informative toast notifications to the user
